+++ equinox/urls.py
@@ -49,6 +49,7 @@ urlpatterns = [
                   path('', include('start.urls')),  # Equinox Start Page URLS
                   path('reporting/', include('reporting.urls')),  # Results Explorer URLS
                   path('policy/', include('policy.urls')),  # Results Explorer URLS
+                  path('reference/', include('reference.urls')),  # Results Explorer URLS
                   path('risk/', include('risk.urls')),  # Risk Analysis URLS
                   path('admin/doc/', include('django.contrib.admindocs.urls')),  #
                   path('admin/', admin.site.urls),  # Equinox Admin URL's