+++ CHANGELOG.rst
@@ -3,11 +3,13 @@ ChangeLog
 
 This is the Changelog for the overall Equinox platform and its API changes. Individual apps may have their own more detailed changelogs within the app subdirectory
 
-v0.6 (XX-06-2022)
+v0.6 (15-06-2022)
 -----------------
 * Enhancements: cpv / cpa mapping
 * Documentation: expand to cover app views
 * Dependencies: added scipy, bumped to latest versions
+* Summary Statistics Reporting and Visualization
+* Adopt behave BDD testing framework
 
 v0.5 (25-05-2022)
 -----------------