+++ reference/CHANGELOG.rst
@@ -1,6 +1,10 @@
 ChangeLog
 ===========================
 
+v0.4.0 (15-06-2022)
+-------------------
+* CPV data
+
 v0.3.0 (21-04-2022)
 -------------------
 * Earth Day 2022 release (starting separate changelog for this app)