+++ reporting/CHANGELOG.rst
@@ -1,6 +1,11 @@
 ChangeLog
 ===========================
 
+v0.4.0 (14-06-2022)
+-------------------
+* Summary Statistics Model
+* Country / Sector visualization
+
 v0.3.0 (21-04-2022)
 -------------------
 * Earth Day 2022 release